Use tracked URL in L'etudiant mission sync (#272)
* Use getMissionTrackedApplicationUrl

* Add some tests for other utils functions

* Use helper in controllers
